+++ build-system/compile/generate/metadata/shared-bento-symbols.js
